requirements.txt
asgiref==3.7.2
coverage==7.4.0
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
